requirements.txt
aiohttp~=3.9.1
coloredlogs~=15.0.1
APScheduler~=3.10.4
python-dotenv~=1.0.0